tvl: abstract TTControlScroll
